finished deleting and cleaning using visitor walk trees

+++ models/nio/DeleteFileVisitor.cfc
@@ -2,10 +2,8 @@
  * Copyright Since 2005 ColdBox Framework by Luis Majano and Ortus Solutions, Corp
  * www.ortussolutions.com
  * ---
- * This is a CFML implementation of a Java nio FileVisitor class
+ * This file visitor only recursively deletes files
  *
- * @see    https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/nio/file/FileVisitor.html
- * @see    https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/nio/file/FileVisitResult.html
  * @author Luis Majano <lmajano@ortussolutions.com>, Grant Copley <gcopley@ortussolutions.com>
  */
 component extends="FileVisitor" {
@@ -21,7 +19,9 @@ component extends="FileVisitor" {
 	 * @throws IOException - if an I/O error occurs
 	 */
 	function visitFile( file, attrs ){
-		variables.jFiles.delete( arguments.file );
+		if ( variables.jFiles.isRegularFile( arguments.file, [] ) ) {
+			variables.jFiles.delete( arguments.file );
+		}
 		return variables.jFileVisitResult.CONTINUE;
 	}
